Converted BaseView (and other views) constructors from (int,int) to (float,float) Added DebugView Fixed blend/gamma input change errors

+++ src/BaseView.cpp
@@ -9,7 +9,7 @@
 
 #include "BaseView.h"
 
-BaseView::BaseView(int width, int height) {
+BaseView::BaseView(float width, float height) {
 	/* save parameters */
 	_viewWidth = width;
 	_viewHeight = height;
